LayoutTests/imported/w3c/web-platform-tests/shadow-dom/reference-target/tentative/form-expected.txt



FAIL Reference target works with form attribute. assert_equals: The .elements property should have 2 elements. expected 2 but got 1
FAIL Reference target works with form attribute via options. assert_equals: The .elements property should have 2 elements. expected 2 but got 1
PASS Reference target works with setAttribute('form')
FAIL Reference target works with form-associated custom element. realForm.moveBefore is not a function. (In 'realForm.moveBefore(customElementInShadow, input)', 'realForm.moveBefore' is undefined)
FAIL Reference target works with nested shadow trees. assert_equals: The .elements property should have 3 elements. expected 3 but got 2
FAIL Form association is updated when form is inserted in or removed from shadow DOM, including in nested shadow DOM assert_equals: The first element should be the reset button outside the shadow root. expected Element node <button id="reset-button-6" type="reset" form="fancy-form... but got Element node <input id="input-in-shadow" value="default value" form="r...
FAIL Form association works for image inputs, which aren't in the elements collection assert_equals: expected Element node <fancy-form-7 id="fancy-form-7">

  </fancy-form-7> but got null
FAIL Changing the reference target of a nested shadow root sets form association assert_equals: Form is outer-fancy-form after reference target change expected Element node <fancy-form-8 id="fancy-form-8">

  </fancy-form-8> but got null
